+++ app/src/test/java/dev/barna/calm/CardVibrancyLevelTest.kt
@@ -8,7 +8,7 @@ class CardVibrancyLevelTest {
     @Test
     fun defaultVibrancyLeavesAlphaUnchanged() {
         // vibrancy=50: baseAlpha * 2 * (100-50) / 100 = baseAlpha * 1.0
-        val color = 0x520F0F14.toInt() // CalmTheme.GLASS: alpha=82, r=15, g=15, b=20
+        val color = 0x520F0F14 // CalmTheme.GLASS: alpha=82, r=15, g=15, b=20
         val result = CardVibrancyLevel.applyTo(color, 50)
         assertEquals(82, (result ushr 24) and 0xFF)
     }
@@ -16,7 +16,7 @@ class CardVibrancyLevelTest {
     @Test
     fun zeroVibrancyDoublesAlpha() {
         // vibrancy=0: baseAlpha * 2 * 100 / 100 = baseAlpha * 2, clamped at 255
-        val color = 0x520F0F14.toInt() // alpha=82
+        val color = 0x520F0F14 // alpha=82
         val result = CardVibrancyLevel.applyTo(color, 0)
         assertEquals(164, (result ushr 24) and 0xFF)
     }
@@ -24,7 +24,7 @@ class CardVibrancyLevelTest {
     @Test
     fun maxVibrancyZerosAlpha() {
         // vibrancy=100: baseAlpha * 2 * 0 / 100 = 0
-        val color = 0x520F0F14.toInt()
+        val color = 0x520F0F14
         val result = CardVibrancyLevel.applyTo(color, 100)
         assertEquals(0, (result ushr 24) and 0xFF)
     }
@@ -49,7 +49,7 @@ class CardVibrancyLevelTest {
 
     @Test
     fun outOfRangeVibrancyIsCoerced() {
-        val color = 0x520F0F14.toInt() // alpha=82
+        val color = 0x520F0F14 // alpha=82
         // vibrancy=-10 should be treated as 0: alpha doubled
         val resultLow = CardVibrancyLevel.applyTo(color, -10)
         assertEquals(164, (resultLow ushr 24) and 0xFF)
@@ -61,7 +61,7 @@ class CardVibrancyLevelTest {
     @Test
     fun midpointVibrancy25IncreasesOpacity() {
         // vibrancy=25: baseAlpha * 2 * 75 / 100 = baseAlpha * 1.5
-        val color = 0x640F0F14.toInt() // alpha=100
+        val color = 0x640F0F14 // alpha=100
         val result = CardVibrancyLevel.applyTo(color, 25)
         assertEquals(150, (result ushr 24) and 0xFF)
     }
